+++ org.lflang/src/lib/core/platform/lf_windows_support.h
@@ -85,7 +85,7 @@ typedef long long int _interval_t;
 /**
  * Microstep instant.
  */
-typedef unsigned int _microstep_t;
+typedef unsigned long int _microstep_t;
 
 #define _LF_TIMEOUT ETIMEDOUT
